+++ ostap/fitting/signals.py
@@ -657,7 +657,7 @@ class Needham_pdf(PEAK) :
                                     "c_{0}(%s)" % name  ,
                                     True , 2.5 , 1.5    , 3.5 ) 
         
-        s_minmax = self.sigma.minmax
+        s_minmax = self.sigma.minmax()
         if s_minmax :
             smin, smax = s_minmax
             c1limits = 0.1 * smin , 10 * smax